Modified structure and added search type
